Refactor smolder_createdb to be less MySQL-specific.
The general strategy employed here has been to move away from calling
the 'mysql' binary directly, to routing calls through the DBI interface. 

This way, we can isolate many driver differences back to the SQL files, or the
connect strings, which DBIx::DBH handles for us. 

I've tested that smolder_createdb --dry_run gives a reasonable result for both
SQLite and MySQL.  MySQL supports seems to be working, but I don't have access
to test it completely at the moment. 

./smolder_createdb will still currently fail horribly for SQLite, because I
haven't written the DBIx::DBH::SQLite driver. 

Also, we need to give it an appropriate path name to store the SQLite db file at.

Once that step is done, we can at least run the application and begin to hunt for SQL
statements that need to be refactored to be compatible. 

    Mark